Moved serialize and deserialize to Blueprint
