+++ src/js/game/blueprint.js
@@ -5,6 +5,10 @@ import { Vector } from "../core/vector";
 import { Entity } from "./entity";
 import { ACHIEVEMENTS } from "../platform/achievement_provider";
 import { GameRoot } from "./root";
+import { SerializerInternal } from "../savegame/serializer_internal";
+import { createLogger } from "../core/logging";
+
+const logger = createLogger("blueprint");
 
 export class Blueprint {
     /**
@@ -12,13 +16,7 @@ export class Blueprint {
      */
     constructor(entities) {
         this.entities = entities;
-    }
-
-    /**
-     * Returns array of entities
-     */
-    getEntities() {
-        return this.entities;
+        this.serializer = new SerializerInternal();
     }
 
     /**
@@ -30,6 +28,47 @@ export class Blueprint {
             return "regular";
         }
         return this.entities[0].layer;
+    }
+
+    /**
+     * Serialize
+     */
+    serialize() {
+        return this.serializer.serializeEntityArray(this.entities);
+    }
+
+    /**
+     * Deserialize
+     * @param {GameRoot} root
+     * @param {Object} json
+     * @retruns {Blueprint|void}
+     */
+    static deserialize(root, json) {
+        try {
+            if (typeof json != "object") {
+                return;
+            }
+            if (!Array.isArray(json)) {
+                return;
+            }
+            /** @type {Array<Entity>} */
+            const entityArray = [];
+            for (let i = 0; i < json.length; ++i) {
+                /** @type {Entity?} */
+                const value = json[i];
+                if (value.components == undefined || value.components.StaticMapEntity == undefined) {
+                    return;
+                }
+                const result = new SerializerInternal().deserializeEntity(root, value);
+                if (typeof result === "string") {
+                    throw new Error(result);
+                }
+                entityArray.push(result);
+            }
+            return new Blueprint(entityArray);
+        } catch (e) {
+            logger.error("Invalid blueprint data:", e.message);
+        }
     }
 
     /**
